backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.11.7
qrcode[pil]==8.2
email-validator==2.2.0
requests==2.32.3
